[Anel] update view and update the binding --no-edit

+++ app/src/main/java/com/anelcc/coronavirustrack/ui/countrydetail/CountryDetail.kt
@@ -30,6 +30,6 @@ class CountryDetail {
     var critical: Int? = null
     @SerializedName("casesPerOneMillion")
     @Expose
-    var casesPerOneMillion: Int? = null
+    var casesPerOneMillion: Double? = null
 
 }
